算法第四版 2.1
2.1
选择排序
插入排序
希尔排序

+++ Algorithm_Exercises/Algorithm (4th edition)/Algorithm/StandardLibraries/StopWatch.cs
@@ -3,7 +3,7 @@ namespace StandardLibraries
 {
     public class StopWatch
     {
-        private readonly int start;
+        private readonly long start;
         private static readonly DateTime Jan1st1970 = new DateTime
     (1970, 1, 1, 0, 0, 0, DateTimeKind.Utc);
         public StopWatch()
@@ -13,6 +13,6 @@ namespace StandardLibraries
 
         public double ElapsedTime => (CurrentTimeMillis - start) / 1000.0;
 
-        public static int CurrentTimeMillis => (int)(DateTime.UtcNow - Jan1st1970).TotalMilliseconds;
+        public static long CurrentTimeMillis => (long)(DateTime.UtcNow - Jan1st1970).TotalMilliseconds;
     }
 }
